Squashed commit of the following:

commit 6af36541b11af3b7050d184b5ab22d038f7917c8
Author: Yuduo Wu <yuduo@apple.com>
Date:   Wed Jul 15 13:24:52 2020 -0700

    rdar://62466374 (TF2 Op Test: TF2 seems to error out when datatype=float16 on ElementWiseBinary tests.)

+++ coremltools/converters/mil/testing_utils.py
@@ -108,8 +108,7 @@ def random_gen(
     If allow_duplicate is set to false, it is guaranteed that value generated are all different.
     Default data type is np.float32.
     """
-
-    elem = np.prod(shape)
+    elem = np.prod(shape).astype(np.int)
     ret = []
     for _ in range(elem):
         while True:
